services/coordinator/requirements.txt
docker
kubernetes
Jinja2
pyyaml
ray[tune]
ax-platform
tqdm